feat(sidebars-mutations): moved resolvers into components/Sidebar directory, as there entire logic is located
