+++ examples/shop-with-blog/client/src/clientState.js
@@ -1,4 +1,5 @@
-import { GET_SIDEBAR_STATE } from 'src/components/Sidebar';
+import { openSidebarResolver, closeSidebarResolver } from 'src/components/Sidebar';
+
 /**
  * Defines client-side state resolvers
  */
@@ -63,33 +64,8 @@ export default {
     },
 
     Mutation: {
-      openSidebar: (_, { contentType, side }, { cache }) => {
-        const data = {
-          sidebar: {
-            contentType,
-            side: side || 'right',
-            isOpen: true,
-            __typename: 'SidebarStatus'
-          }
-        };
-
-        cache.writeQuery({ query: GET_SIDEBAR_STATE, data });
-
-        return null;
-      },
-
-      closeSidebar: (_, _variables, { cache }) => {
-        const queryResponse = cache.readQuery({ query: GET_SIDEBAR_STATE });
-        const sidebar = { ...queryResponse.sidebar };
-        sidebar.isOpen = false;
-
-        cache.writeQuery({
-          query: GET_SIDEBAR_STATE,
-          data: { sidebar }
-        });
-
-        return null;
-      }
+      openSidebar: openSidebarResolver,
+      closeSidebar: closeSidebarResolver
     }
   }
 };
