+++ examples/shop-with-blog/client/src/components/Sidebar/index.js
@@ -1,5 +1,7 @@
 export * from './CloseSidebarMutation';
+export * from './closeSidebarResolver';
 export * from './OpenSidebarMutation';
+export * from './openSidebarResolver';
 export * from './SidebarQuery';
 export * from './SidebarContainer';
 export * from './sidebarType';
